# ai/wait_time_module/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-Native Wait-Time & Queue Engine",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/events": {
      "post": {
        "summary": "Event",
        "operationId": "event_api_v1_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/patients/{patient_token}/estimate": {
      "get": {
        "summary": "Patient",
        "operationId": "patient_api_v1_patients__patient_token__estimate_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "patient_token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientEstimate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/queues/{queue_id}/estimates": {
      "get": {
        "summary": "Queue",
        "operationId": "queue_api_v1_queues__queue_id__estimates_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "queue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Queue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/journeys/{journey_id}/estimate": {
      "get": {
        "summary": "Journey",
        "operationId": "journey_api_v1_journeys__journey_id__estimate_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "journey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Journey Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rooms/status": {
      "get": {
        "summary": "Rooms Status",
        "operationId": "rooms_status_api_v1_rooms_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/queues/{queue_id}/summary": {
      "get": {
        "summary": "Queue Summary",
        "operationId": "queue_summary_api_v1_queues__queue_id__summary_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "queue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Queue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/estimates/room-options": {
      "post": {
        "summary": "Room Options",
        "operationId": "room_options_api_v1_estimates_room_options_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomOptionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/estimates/assignment-impact": {
      "post": {
        "summary": "Assignment Impact",
        "operationId": "assignment_impact_api_v1_estimates_assignment_impact_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentImpactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/simulations/scenario": {
      "post": {
        "summary": "Scenario",
        "operationId": "scenario_api_v1_simulations_scenario_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AssignmentImpactRequest": {
        "properties": {
          "patient_token": {
            "type": "string",
            "title": "Patient Token"
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "service_code": {
            "type": "string",
            "title": "Service Code"
          },
          "clinical_priority": {
            "$ref": "#/components/schemas/ClinicalPriority"
          },
          "ready_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ready At"
          },
          "room_id": {
            "type": "string",
            "title": "Room Id"
          },
          "predicted_minutes": {
            "type": "number",
            "minimum": 0.0,
            "title": "Predicted Minutes",
            "default": 12
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "patient_token",
          "task_type",
          "service_code",
          "clinical_priority",
          "ready_at",
          "room_id"
        ],
        "title": "AssignmentImpactRequest"
      },
      "ClinicalPriority": {
        "type": "string",
        "enum": [
          "EMERGENCY",
          "URGENT",
          "NORMAL"
        ],
        "title": "ClinicalPriority"
      },
...
      "Event": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "event_time": {
            "type": "string",
            "format": "date-time",
            "title": "Event Time"
          },
          "journey_id": {
            "type": "string",
            "title": "Journey Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "patient_token": {
            "type": "string",
            "title": "Patient Token"
          },
          "queue_id": {
            "type": "string",
            "title": "Queue Id"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType"
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType",
            "default": "INITIAL_CONSULT"
          },
          "clinical_priority": {
            "$ref": "#/components/schemas/ClinicalPriority",
            "default": "NORMAL"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "based_on_estimate_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Based On Estimate Version"
          },
          "actor_id": {
            "type": "string",
            "title": "Actor Id",
            "default": "unknown"
          },
          "actor_type": {
            "type": "string",
            "title": "Actor Type",
            "default": "SERVICE"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "event_time",
          "journey_id",
          "task_id",
          "patient_token",
          "queue_id",
          "event_type"
        ],
        "title": "Event"
      },
      "EventResponse": {
        "properties": {
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          },
          "duplicate": {
            "type": "boolean",
            "title": "Duplicate"
          },
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "estimate_version": {
            "type": "integer",
            "title": "Estimate Version"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "OK"
          }
        },
        "type": "object",
        "required": [
          "accepted",
          "duplicate",
          "event_id",
          "estimate_version"
        ],
        "title": "EventResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskType": {
        "type": "string",
        "enum": [
          "INITIAL_CONSULT",
          "DIAGNOSTIC_SERVICE",
          "RETURN_REVIEW",
          "FOLLOW_UP_CONSULT"
        ],
        "title": "TaskType"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}